feat: add feedback, login, feedback, like

+++ app/src/main/java/com/example/rent_a_car_app/room/CarsData.kt
@@ -7,7 +7,7 @@ import androidx.room.PrimaryKey
 data class CarsData(
     @PrimaryKey(autoGenerate = true) var id: Int,
     var name: String,
-    var price: Double,
+    var price: Int,
     var description: String,
     val img: String,
     var favorite: Boolean
